Adds price picker component
Introduces a reusable price picker component for selecting dollar and cent values.

Replaces number inputs with the new component in the shopping list manager for setting prices.
The price picker is implemented using wheel pickers and allows for selecting dollar and cent amounts.
A modal is now used to set both new and existing item prices.